+++ internal/app/models/pasture/water_chemical_hazard.go
@@ -4,24 +4,24 @@ import "gorm.io/gorm"
 
 type PastureWaterChemicalHazard struct {
 	gorm.Model
-	Fluoride        float32 `json:"fluoride"`         // 氟化物
-	Cyanide         float32 `json:"cyanide"`          // 氰化物
-	Chloride        float32 `json:"chloride"`         // 氰化物
-	Nitrate         float32 `json:"nitrate"`          // 硝酸盐
-	Sulfate         float32 `json:"sulfate"`          // 硫酸盐
-	Sixsixsix       float32 `json:"sixsixsix"`        // 六六六
-	DDT             float32 `json:"ddt"`              // 滴滴涕
-	AmmoniaNitrogen float32 `json:"ammonia_nitrogen"` // 氨氮
+	Fluoride        float64 `json:"fluoride"`         // 氟化物
+	Cyanide         float64 `json:"cyanide"`          // 氰化物
+	Chloride        float64 `json:"chloride"`         // 氰化物
+	Nitrate         float64 `json:"nitrate"`          // 硝酸盐
+	Sulfate         float64 `json:"sulfate"`          // 硫酸盐
+	Sixsixsix       float64 `json:"sixsixsix"`        // 六六六
+	DDT             float64 `json:"ddt"`              // 滴滴涕
+	AmmoniaNitrogen float64 `json:"ammonia_nitrogen"` // 氨氮
 	PastureWaterID  uint    `json:"pasture_water_id"`
 }
 
 type PastureWaterChemicalHazardData struct {
-	Fluoride        float32 `json:"fluoride"`         // 氟化物
-	Cyanide         float32 `json:"cyanide"`          // 氰化物
-	Chloride        float32 `json:"chloride"`         // 氰化物
-	Nitrate         float32 `json:"nitrate"`          // 硝酸盐
-	Sulfate         float32 `json:"sulfate"`          // 硫酸盐
-	Sixsixsix       float32 `json:"sixsixsix"`        // 六六六
-	DDT             float32 `json:"ddt"`              // 滴滴涕
-	AmmoniaNitrogen float32 `json:"ammonia_nitrogen"` // 氨氮
+	Fluoride        float64 `json:"fluoride"`         // 氟化物
+	Cyanide         float64 `json:"cyanide"`          // 氰化物
+	Chloride        float64 `json:"chloride"`         // 氰化物
+	Nitrate         float64 `json:"nitrate"`          // 硝酸盐
+	Sulfate         float64 `json:"sulfate"`          // 硫酸盐
+	Sixsixsix       float64 `json:"sixsixsix"`        // 六六六
+	DDT             float64 `json:"ddt"`              // 滴滴涕
+	AmmoniaNitrogen float64 `json:"ammonia_nitrogen"` // 氨氮
 }
